memlinkto_project/requirements.txt
Django==4.1.6
django-bootstrap5==22.2
django-allauth==0.52.0
django-modern-rpc==1.0.1
django-hCaptcha==0.2.0